requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
nltk>=3.6.2
newspaper3k>=0.2.8
scikit-learn>=0.24.2
python-multipart
Jinja2>=2.11.3
lxml_html_clean>=1.0.0